requirements.txt
django-ckeditor
pillow
pytils
easy-thumbnails==2.7